requirements.txt
openai>=1.0.0
python-dotenv>=0.21.0
tqdm>=4.64.0
